requirements.txt
# Requirements for Personal Productivity Assistant

rich
prompt_toolkit
python-dateutil
